+++ packages/vuelidate/src/index.js
@@ -1,12 +1,23 @@
-import { watch, computed, getCurrentInstance, inject, onBeforeMount, onBeforeUnmount, provide, isRef, ref, reactive, isProxy } from 'vue-demi'
-import { isFunction, unwrap } from './utils'
+import { watch, computed, getCurrentInstance, inject, onBeforeMount, onBeforeUnmount, provide, isRef, ref, reactive } from 'vue-demi'
+import { isFunction, unwrap, isProxy } from './utils'
 import { setValidations } from './core'
 import ResultsStorage from './storage'
 
 const VuelidateInjectChildResults = Symbol('vuelidate#injectChiildResults')
 const VuelidateRemoveChildResults = Symbol('vuelidate#removeChiildResults')
 
-function nestedValidations () {
+export const CollectFlag = {
+  COLLECT_ALL: true,
+  COLLECT_NONE: false
+}
+
+/**
+ * Create helpers to collect validation state from child components
+ * @param {Object} params
+ * @param {String | Number} params.$scope - Parent component scope
+ * @return {{sendValidationResultsToParent: function, childResults: ComputedRef<Object>, removeValidationResultsFromParent: function}}
+ */
+function nestedValidations ({ $scope }) {
   const childResultsRaw = {}
   const childResultsKeys = ref([])
   const childResults = computed(() => childResultsKeys.value.reduce((results, key) => {
@@ -17,9 +28,20 @@ function nestedValidations () {
   /**
    * Allows children to send validation data up to their parent.
    * @param {Object} results - the results
-   * @param {String} key - the registeredAs key
+   * @param {Object} args
+   * @param {String} args.$registerAs - the $registeredAs key
+   * @param {String | Number} args.$scope - the $scope key
    */
-  function injectChildResultsIntoParent (results, key) {
+  function injectChildResultsIntoParent (results, { $registerAs: key, $scope: childScope, $stopPropagation }) {
+    if (
+      $stopPropagation ||
+      $scope === CollectFlag.COLLECT_NONE ||
+      childScope === CollectFlag.COLLECT_NONE ||
+      (
+        $scope !== CollectFlag.COLLECT_ALL &&
+        $scope !== childScope
+      )
+    ) return
     childResultsRaw[key] = results
     childResultsKeys.value.push(key)
   }
@@ -35,6 +57,7 @@ function nestedValidations () {
     delete childResultsRaw[key]
   }
 
+  // inject the `injectChildResultsIntoParent` method, into the current scope
   const sendValidationResultsToParent = inject(VuelidateInjectChildResults, () => {})
   // provide to all of it's children the send results to parent function
   provide(VuelidateInjectChildResults, injectChildResultsIntoParent)
@@ -47,15 +70,29 @@ function nestedValidations () {
 }
 
 /**
+ * @typedef GlobalConfig
+ * @property {String} [$registerAs] - Config Object
+ * @property {String | Number | Symbol} [$scope] - A scope to limit child component registration
+ * @property {Boolean} [$stopPropagation] - Tells a Vue component to stop sending it's results up to the parent
+ */
+
+/**
  * Composition API compatible Vuelidate
  * Use inside the `setup` lifecycle hook
- * @param {Object|null} validations - Validations Object
- * @param {Object} state - State object
- * @param {String} globalConfig - Config Object
+ * @param {Object | GlobalConfig} validations - Validations Object or the globalConfig.
+ * @param {Object} [state] - State object - required if `validations` is a validation object.
+ * @param {GlobalConfig} [globalConfig] - Config Object
  * @return {UnwrapRef<*>}
  */
 export function useVuelidate (validations, state, globalConfig = {}) {
-  let { $registerAs } = globalConfig
+  // if we pass only one argument, its most probably the globalConfig.
+  // This use case is so parents can just collect results of child forms.
+  if (arguments.length === 1) {
+    globalConfig = validations
+    validations = undefined
+    state = undefined
+  }
+  let { $registerAs, $scope = CollectFlag.COLLECT_ALL, $stopPropagation } = globalConfig
 
   const instance = getCurrentInstance()
 
@@ -70,7 +107,7 @@ export function useVuelidate (validations, state, globalConfig = {}) {
   const validationResults = ref({})
   const resultsCache = new ResultsStorage()
 
-  const { childResults, sendValidationResultsToParent, removeValidationResultsFromParent } = nestedValidations()
+  const { childResults, sendValidationResultsToParent, removeValidationResultsFromParent } = nestedValidations({ $scope, $stopPropagation })
 
   // Options API
   if (!validations && instance.type.validations) {
@@ -127,7 +164,7 @@ export function useVuelidate (validations, state, globalConfig = {}) {
   }
 
   // send all the data to the parent when the function is invoked inside setup.
-  sendValidationResultsToParent(validationResults, $registerAs)
+  sendValidationResultsToParent(validationResults, { $registerAs, $scope, $stopPropagation })
   // before this component is destroyed, remove all the data from the parent.
   onBeforeUnmount(() => removeValidationResultsFromParent($registerAs))