+++ packages/vuelidate/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.11](https://github.com/vuelidate/vuelidate/compare/@vuelidate/core@2.0.0-alpha.10...@vuelidate/core@2.0.0-alpha.11) (2021-02-10)
+
+
+### Bug Fixes
+
+* **core:** remove isProxy and implement own ([95706a3](https://github.com/vuelidate/vuelidate/commit/95706a3ee667ae8d01b37172868360277da86c7f))
+* **core:** remove isProxy and implement own in index.js ([376c054](https://github.com/vuelidate/vuelidate/commit/376c054844fcde8529e572a9bd1690a38621f5f7))
+
+
+
+
+
+# [2.0.0-alpha.10](https://github.com/vuelidate/vuelidate/compare/@vuelidate/core@2.0.0-alpha.9...@vuelidate/core@2.0.0-alpha.10) (2021-02-07)
+
+
+### Bug Fixes
+
+* **core:** allow $autoDirty on the root, when passed a plain object as state ([#796](https://github.com/vuelidate/vuelidate/issues/796)) ([4d49bb2](https://github.com/vuelidate/vuelidate/commit/4d49bb2060bfa07f8e190489e89cb5240b37311d))
+
+
+
+
+
+# [2.0.0-alpha.9](https://github.com/vuelidate/vuelidate/compare/@vuelidate/core@2.0.0-alpha.8...@vuelidate/core@2.0.0-alpha.9) (2021-02-04)
+
+
+### Features
+
+* add global configs, remove mixin, add silentErrors, lazy mode  ([#790](https://github.com/vuelidate/vuelidate/issues/790)) ([22cd7c5](https://github.com/vuelidate/vuelidate/commit/22cd7c5ae5a0c5c2e4a021dc082509b3be3f5141)), closes [#670](https://github.com/vuelidate/vuelidate/issues/670)
+* allow scoping of component validations, fixes: [#719](https://github.com/vuelidate/vuelidate/issues/719) ([#791](https://github.com/vuelidate/vuelidate/issues/791)) ([d7a8797](https://github.com/vuelidate/vuelidate/commit/d7a87976f3a1a27914c406462e09df69ccb653c8))
+
+
+### BREAKING CHANGES
+
+* Validations are no longer lazy by default, you need to specify `$lazy: true` to each, or at the top using the globalConfig.
+
+
+
+
+
 # [2.0.0-alpha.8](https://github.com/vuelidate/vuelidate/compare/@vuelidate/core@2.0.0-alpha.7...@vuelidate/core@2.0.0-alpha.8) (2020-11-19)